# Conflicts:
#	src/main/java/org/blank/projectmanagementsystem/repository/ArchitectureRepository.java
#	src/main/java/org/blank/projectmanagementsystem/repository/DeliverableRepository.java
#	src/main/java/org/blank/projectmanagementsystem/repository/ProjectRepository.java
#	src/main/java/org/blank/projectmanagementsystem/repository/SystemOutlineRepository.java
#	src/main/resources/static/css/style.css

+++ src/main/java/org/blank/projectmanagementsystem/repository/SystemOutlineRepository.java
@@ -1,5 +1,4 @@
 package org.blank.projectmanagementsystem.repository;
-
 import org.blank.projectmanagementsystem.domain.entity.Department;
 import org.blank.projectmanagementsystem.domain.entity.SystemOutline;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -7,6 +6,6 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.Optional;
 
-public interface SystemOutlineRepository extends JpaRepository<SystemOutline, Long>{
+public interface SystemOutlineRepository extends JpaRepository<SystemOutline, Integer>{
     SystemOutline findByName(String systemOutlineName);
 }
